Respond to Review of PR 3827
