+++ dds/DCPS/XTypes/DynamicTypeSupport.h
@@ -3,8 +3,8 @@
  * See: http://www.opendds.org/license.html
  */
 
-#ifndef OPENDDS_DCPS_XTYPES_DYNAMIC_TYPE_SUPPORT_IMPL_H
-#define OPENDDS_DCPS_XTYPES_DYNAMIC_TYPE_SUPPORT_IMPL_H
+#ifndef OPENDDS_DCPS_XTYPES_DYNAMIC_TYPE_SUPPORT_H
+#define OPENDDS_DCPS_XTYPES_DYNAMIC_TYPE_SUPPORT_H
 
 #ifndef OPENDDS_SAFETY_PROFILE
 #  include <dds/DCPS/Sample.h>
@@ -287,4 +287,5 @@ TAO_END_VERSIONED_NAMESPACE_DECL
 #  endif
 
 #endif // OPENDDS_SAFETY_PROFILE
-#endif // OPENDDS_DCPS_XTYPES_DYNAMIC_TYPE_SUPPORT_IMPL_H
+
+#endif // OPENDDS_DCPS_XTYPES_DYNAMIC_TYPE_SUPPORT_H
